refactor: implement quint verification and clean up stale comments
- Add verify_invariant() and verify_temporal() methods to QuintEvaluator
  for running actual Quint CLI model checking commands
- Update verify_capability_invariants() to use new evaluator methods
  with proper error aggregation and reporting
- Add InvariantVerificationResult and TemporalVerificationResult types

- Remove stale "moved to" comments across crates:
  - aura-core: epochs, identifiers, relationships type comments
  - aura-transport: amp module doc
  - aura-signature: removed DKD message comments
  - aura-effects: mock handler location comments
  - aura-protocol: handler and effect module notes
  - aura-journal: type and effect comments
  - aura-terminal: redundant architecture doc lines

+++ crates/aura-harness/tests/phase2_routing_replay.rs
@@ -92,7 +92,7 @@ fn replay_runner_reexecutes_recorded_actions_without_llm() {
         Err(error) => panic!("replay execute failed: {error}"),
     };
 
-    assert_eq!(outcome.actions_executed, bundle.actions.len());
+    assert_eq!(outcome.actions_executed, bundle.actions.len() as u64);
     assert_eq!(outcome.mismatches, 0);
 }
